requirements.txt
# Core data processing libraries
polars>=0.19.19
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine learning libraries
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
hyperopt>=0.2.7
shap>=0.43.0

# Visualization libraries
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Utility libraries
pyyaml>=6.0.0
joblib>=1.3.0
umap-learn>=0.5.4

# File format support
pyarrow>=14.0.0

# API and deployment
fastapi>=0.104.0
uvicorn>=0.23.2
gunicorn>=21.2.0
pydantic>=2.4.2

# Monitoring and scheduling
schedule>=1.2.0
prometheus-client>=0.17.1

# Dashboard
streamlit>=1.28.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
sphinx>=7.2.6
mkdocs>=1.5.3 